requirements.txt
ultralytics>=8.0.0
streamlit>=1.32.0
anthropic>=0.25.0
Pillow>=9.0.0
datasets>=2.18.0
